Wiring input_pc_ready to first FIFO
This let us keep our original switching logic, but we need the arbiter
on the first input to always grant the input_pc_and_cc_id: this is
needed because we are artificially granting it using the ready signal
of the first fifo, hence we do not have any way to stall it.

This was done to avoid a deadlock on the switch not sending a valid signal
until it receives a ready signal, and the arbiter not sending a ready signal until it
receives a valid signal - hence we had a deadlock.